# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maya Import Export Logistic API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      parameters:
        - in: query
          name: role
          required: false
          schema:
            type: string
            enum: [admin, staff, customer]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a staff or customer account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/permissions:
    patch:
      operationId: updateUserPermissions
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionsBody"
      responses:
        "200":
          description: Updated permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permissions"
  /users/{id}/reset-password:
    post:
      operationId: adminResetUserPassword
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResetPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ----- Shipments -----
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Role:
      type: string
      enum: [admin, staff, customer]
...
    Permissions:
      type: object
      properties:
        canManageShipments:
          type: boolean
        canManageCustomers:
          type: boolean
        canGenerateInvoice:
          type: boolean
      required: [canManageShipments, canManageCustomers, canGenerateInvoice]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
        permissions:
          $ref: "#/components/schemas/Permissions"
      required: [id, name, email, phone, role, createdAt, permissions]
...
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/Role"
        permissions:
          $ref: "#/components/schemas/Permissions"
      required: [name, email, phone, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    PermissionsBody:
      type: object
      properties:
        canManageShipments:
          type: boolean
        canManageCustomers:
          type: boolean
        canGenerateInvoice:
          type: boolean
      required: [canManageShipments, canManageCustomers, canGenerateInvoice]
    AdminResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
          minLength: 6
      required: [newPassword]
...