# lib/api-zod/src/generated/types/adminResetPasswordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maya Import Export Logistic API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminResetPasswordBody {
  /** @minLength 6 */
  newPassword: string;
}